+++ src/com/intellij/idea/plugin/hybris/project/configurators/impl/DefaultPostImportConfigurator.kt
@@ -26,7 +26,6 @@ import com.intellij.idea.plugin.hybris.project.configurators.PostImportConfigura
 import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
 import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptor
 import com.intellij.idea.plugin.hybris.project.descriptors.impl.MavenModuleDescriptor
-import com.intellij.idea.plugin.hybris.project.utils.PluginCommon
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
@@ -58,10 +57,8 @@ class DefaultPostImportConfigurator(val project: Project) : PostImportConfigurat
         val configuratorFactory = ApplicationManager.getApplication().getService(ConfiguratorFactory::class.java)
 
         try {
-            if (PluginCommon.isPluginActive(PluginCommon.ANT_SUPPORT_PLUGIN_ID)) {
-                configuratorFactory.antConfigurator
-                    ?.configure(hybrisProjectDescriptor, allHybrisModules, project)
-            }
+            configuratorFactory.antConfigurator
+                ?.configure(hybrisProjectDescriptor, allHybrisModules, project)
         } catch (e: Exception) {
             LOG.error("Can not configure Ant due to an error.", e)
         }